refactor: Move AWS Instance Profile crawler to separate file per @nfx
